requirements.txt
# chromadb==0.6.3
# fastapi==0.115.11
# google-ai-generativelanguage==0.6.16
# google-api-core==2.24.1
# google-auth==2.38.0
# groq==0.18.0
# huggingface-hub==0.29.1
# kubernetes==32.0.1
# langchain==0.3.19
# langchain-community==0.3.18
# langchain-core==0.3.40
# langgraph==0.3.2
# langsmith==0.3.11
# openai==1.65.2
# pydantic==2.10.6
# tiktoken==0.9.0
# uvicorn
# langchain_google_genai
# psycopg2
bs4==0.0.2
chromadb==0.6.3
grpcio-status==1.70.0
httptools==0.6.4
ipykernel==6.29.5
langchain-community==0.3.18
langchain-google-genai==2.0.11
langchain-groq==0.2.4
langchain-openai==0.3.7
langchain-postgres==0.0.13
langchainhub==0.1.21
langgraph==0.3.2
pip==23.2.1
psycopg2-binary==2.9.10
setuptools==65.5.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==15.0
fastapi==0.115.11
